Add user purchase feature with Create and get all filtering
Introduces user purchase management including entity, controller, service, repository, DTOs, and filtering specification. Adds error codes for purchase-related errors and updates Plan entity to use Long for coin. Implements purchase creation, validation, and paginated/filterable retrieval of user purchases.

+++ backend/src/main/java/com/mymatch/entity/Plan.java
@@ -29,7 +29,7 @@ public class Plan extends AbstractAuditingEntity {
     @Column(columnDefinition = "TEXT")
     String description;
 
-    Double coin; // coin
+    Long coin; // coin
 
     @Builder.Default
     @Column(nullable = false)
